Move config validation from ringpop to membership
