+++ common/membership/ringpop/factory_test.go
@@ -39,6 +39,7 @@ import (
 	"go.temporal.io/server/common/config"
 	"go.temporal.io/server/common/dynamicconfig"
 	"go.temporal.io/server/common/log"
+	"go.temporal.io/server/common/membership"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/primitives"
 	"go.temporal.io/server/common/rpc/encryption"
@@ -137,19 +138,11 @@ func (s *RingpopSuite) TestHostsMode() {
 	s.Nil(err)
 	s.Equal("1.2.3.4", cfg.BroadcastAddress)
 	s.Equal(time.Second*30, cfg.MaxJoinDuration)
-	err = ValidateConfig(&cfg)
+	err = membership.ValidateConfig(&cfg)
 	s.Nil(err)
 	f, err := newFactory(&cfg, "test", nil, log.NewNoopLogger(), nil, nil, nil, nil)
 	s.Nil(err)
 	s.NotNil(f)
-}
-
-func (s *RingpopSuite) TestInvalidConfig() {
-	var cfg config.Membership
-	cfg.MaxJoinDuration = time.Minute
-	s.NoError(ValidateConfig(&cfg))
-	cfg.BroadcastAddress = "sjhdfskdjhf"
-	s.Error(ValidateConfig(&cfg))
 }
 
 func getHostsConfig() string {
