+++ temporal/fx.go
@@ -52,7 +52,7 @@ import (
 	"go.temporal.io/server/common/headers"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
-	"go.temporal.io/server/common/membership/ringpop"
+	"go.temporal.io/server/common/membership"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/cassandra"
@@ -169,7 +169,7 @@ func ServerOptionsProvider(opts []ServerOption) (serverOptionsProvider, error) {
 		return serverOptionsProvider{}, err
 	}
 
-	err = ringpop.ValidateConfig(&so.config.Global.Membership)
+	err = membership.ValidateConfig(&so.config.Global.Membership)
 	if err != nil {
 		return serverOptionsProvider{}, fmt.Errorf("ringpop config validation error: %w", err)
 	}
